+++ .gitignore
@@ -26,3 +26,5 @@ build/Release
 # https://www.npmjs.org/doc/misc/npm-faq.html#should-i-check-my-node_modules-folder-into-git
 node_modules
 *.idea
+
+package-lock.json
